clip-service/requirements.txt
torch
transformers
Pillow
fastapi
uvicorn
httpx
